TT2-291 refactor tests and permissions
