+++ audit/lambda/s3Services/getS3Object.ts
